Minor improvement: use exceptions. Fix tests
